\name{hexmarkers.overdisp}
\docType{data}
\alias{hexmarkers.overdisp}
\title{Simulated overdispersed autopolyploid dominant markers from 200 hexaploid individuals}
\description{

  These data are simulated as 500 markers for 200 \dQuote{auto--hexaploid
    individuals} exhibiting overdispersion with the parameter shape1 =
  25. The underlying percentages of single double and triple dose
  markers are 70\%, 20\% and 10\%, respectively.
}
\usage{hexmarkers.overdisp}
\format{An object of S3 class \code{sim.autoMarkers} containing 500
  simulated dominant markers for 200 auto--hexaploid individuals.}
\references{
    Haldane, J B S. 1930. Theoretical genetics of autopolyploids.
  \emph{Journal of Genetics} 22: 359-372.
  
  Baker P, Jackson P, and Aitken K. 2010. Bayesian estimation of marker
  dosage in sugarcane and other autopolyploids. \emph{TAG Theoretical and
    Applied Genetics} 120 (8): 1653-1672.
}
\keyword{datasets}
