\name{hexmarkers}
\docType{data}
\alias{hexmarkers}
\title{Simulated autopolyploid dominant markers from 200 hexaploid individuals}
\description{These data were simulated as 500 markers for 200
  \dQuote{auto--hexaploid individuals} exhibiting no
  overdispersion. The underlying percentages of single double and
  triple dose markers are 70\%, 20\% and 10\%, respectively.
}
\usage{hexmarkers}
\format{An object of S3 class \code{sim.autoMarkers} containing 500
  simulated dominant markers for 200 auto--hexaploid individuals.}
%%\source{}
\references{
  Haldane, J B S. 1930. Theoretical genetics of autopolyploids.
  \emph{Journal of Genetics} 22: 359-372.
  
  Baker P, Jackson P, and Aitken K. 2010. Bayesian estimation of marker
  dosage in sugarcane and other autopolyploids. \emph{TAG Theoretical and
    Applied Genetics} 120 (8): 1653-1672.
}
\keyword{datasets}
